RW-2329 move vocab api call to fix errors on creation of review
